+++ recipes/views/__init__.py
@@ -1,1 +1,5 @@
-from views import *
+from .ingridientHandler import *
+from .quantityHandler import *
+from .recipeHandler import *
+from .scrapingHandler import *
+from .shoppingListHandler import *
